requirementsNOW.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.12.0
cryptography==41.0.7
deprecation==2.1.0
fastapi==0.115.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
postgrest==2.24.0
prometheus_client==0.21.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.3
realtime==2.24.0
requests==2.32.5
sentry-sdk==2.17.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==2.24.0
StrEnum==0.4.15
structlog==24.4.0
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets>=13.0
yarl==1.22.0
